+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [2.10 dev-090925 avx]
+### Changed
+- Updated engine id name to "Wordfish v2.10 dev-090925 avx".
+- Strengthened king safety evaluation and extended search on exposed enemy kings.
+
 ## [2.0 dev-060925]
 ### Changed
 - Renamed engine to Wordfish 2.0 dev-060925 avx.